feat(front-office): graduate front office generator (rename + DI parity)
Drops the `stub` prefix from the front-office generator and aligns its
dependency-injection shape with the players, coaches, and scouts
generators. The behavior change is intentionally small here: the
`front_office_staff` schema currently exposes only `firstName` /
`lastName` beyond league/team identifiers, so name variance is the only
thing the generator can meaningfully produce. The graduation is still
worth shipping because:

  1. The `stub` prefix implied a placeholder slated for replacement when
     the bottleneck is actually the schema, not this file.
  2. Wiring the shared `NameGenerator` dependency on the same option
     shape used elsewhere means the eventual schema expansion (role,
     title, age, contract, philosophy) becomes a localized edit instead
     of an interface change.

A backlog entry is added recording the schema expansion needed to put
realistic GM / President of Football Ops / Capologist titles on the
generated rows. When that schema lands, this is the natural place to
add tier-band distributions on the same pattern coaches and scouts now
use.

Naming: `git mv` `stub-front-office-generator.{ts,test.ts}` ->
`front-office-generator.{ts,test.ts}`, factory renamed
(`createStubFrontOfficeGenerator` -> `createFrontOfficeGenerator`),
options type renamed (`StubFrontOfficeGeneratorOptions` ->
`FrontOfficeGeneratorOptions`).

Also fixes a stale import in `depth-chart.publisher.test.ts` that landed
on main between #162's merge and this branch — it still pointed at
`stub-players-generator.ts`. The PR-C branch carries the same fix; this
PR carries it too in case PR-C lands later.
