+++ server/features/front-office/mod.ts
@@ -1,3 +1,3 @@
-export { createStubFrontOfficeGenerator } from "./stub-front-office-generator.ts";
+export { createFrontOfficeGenerator } from "./front-office-generator.ts";
 export { createFrontOfficeService } from "./front-office.service.ts";
 export type { FrontOfficeService } from "./front-office.service.interface.ts";
